pyproject.toml
[tool.black]
line-length = 240

[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[tomli]>=6.3"]
build-backend = "setuptools.build_meta"

[project]
name = "easy_sae"
version = "0.1.0"
authors = [
    { name = "Kaichen, Zhang", email = "kaichenzhang358@outlook.com" },
]
description = "A python template"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "peft",
    "transformers",
    "datasets",
    "torchvision",
    "accelerate",
    "deepspeed",
]

[project.optional-dependencies]
opt = [
]

[project.scripts]
easy-sae-train = "easy_sae.launch.train:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = [
    "test"
]

[tool.wheel]
exclude = [
    "test"
]
